\name{nlm_optimize_r}
\alias{nlm_optimize_r}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
NLM optimization
}
\description{
The Non-Linear Minimization function is used by the \code{\link{polynomial}} separator. This function should not be called manually.}


\seealso{
\code{\link{polynomial}}

\code{\link{nlm}}
}
